SONARJAVA-6473 Extract jump continuation helper to CFGUtils
